Fix float values in reinforcement, add gnd and test type to ReinforcementType

+++ core/src/main/java/com/rettichlp/unicacityaddon/commands/faction/ReinforcementCommand.java
@@ -64,9 +64,9 @@ public class ReinforcementCommand extends Command {
         }
 
         FloatVector3 position = p.getPosition();
-        float posX = position.getX();
-        float posY = position.getY();
-        float posZ = position.getZ();
+        int posX = (int) position.getX();
+        int posY = (int) position.getY();
+        int posZ = (int) position.getZ();
 
         if (firstType.getMessage() != null)
             p.sendServerMessage(chatType.getChatCommand() + " " + firstType.getMessage());
